refactor: split plugin-manager utils into focused modules
Converts the 686-line views/services/plugin-manager/utils.tsx into sibling
modules: types (Plugin/NpmConfig/BundlePluginMeta + guards) / npm-utils
(install, remove, repairDep, tarball lookup, mirrors, getNpmConfig) /
plugin-i18n (updateI18n) / read-plugin (readPlugin + bundle-plugin JSX) /
lifecycle (enable/disable/unload, window creation, cache clearing,
notifyFailed). utils.ts remains as a pure barrel with the identical
public surface, so './utils' and 'views/services/plugin-manager/utils'
importers (poi and third-party plugins) are unaffected; the .tsx -> .ts
extension change is invisible to extension-less specifiers.

Pure move. repairDep lives in npm-utils (not lifecycle) because
removePackage -> repairDep -> installPackage would otherwise form an
import cycle. The remote.require singletons move with their consumers
(windowManager -> lifecycle, lib/utils error logger -> read-plugin).
